Sample application and instructions in README
A sample, iteractive aplication in src/client/host.html that exchanges data with external application running in iframe (application.html) and uploads it to the server.

It is also running tests to help determine supported platforms.

Also readme is rewriten. It contains instructions and rationale. Probably leaves a lot to be improved though :)